scripts/pageindex/requirements.txt
pymupdf>=1.26.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
anthropic>=0.40.0
